toolkit2: update in config
git-svn-id: http://svn.terracotta.org/svn/ehcache@6036 b9324663-ca0f-0410-8574-be9b3887307d

+++ branches/ehcache_toolkit2/terracotta/bootstrap/src/main/java/org/terracotta/modules/ehcache/store/ClusteredStoreBackend.java
@@ -225,7 +225,7 @@ public class ClusteredStoreBackend<K, V> implements ToolkitCacheInternal<K, V> {
   }
 
   @Override
-  public void putNoReturn(K key, V value, int createTimeInSecs, int maxTTISeconds, int maxTTLSeconds) {
+  public void putNoReturn(K key, V value, long createTimeInSecs, int maxTTISeconds, int maxTTLSeconds) {
     lock.readLock().lock();
     try {
       activeDelegate.putNoReturn(key, value, createTimeInSecs, maxTTISeconds, maxTTLSeconds);
@@ -235,7 +235,7 @@ public class ClusteredStoreBackend<K, V> implements ToolkitCacheInternal<K, V> {
   }
 
   @Override
-  public V putIfAbsent(K key, V value, int createTimeInSecs, int maxTTISeconds, int maxTTLSeconds) {
+  public V putIfAbsent(K key, V value, long createTimeInSecs, int maxTTISeconds, int maxTTLSeconds) {
     lock.readLock().lock();
     try {
       return activeDelegate.putIfAbsent(key, value, createTimeInSecs, maxTTISeconds, maxTTLSeconds);
