+++ branches/ehcache_toolkit2/terracotta/bootstrap/src/main/java/org/terracotta/modules/ehcache/store/bulkload/BulkLoadToolkitCache.java
@@ -156,13 +156,13 @@ public class BulkLoadToolkitCache<K, V> implements ToolkitCacheInternal<K, V> {
   }
 
   @Override
-  public void putNoReturn(K key, V value, int createTimeInSecs, int customMaxTTISeconds, int customMaxTTLSeconds) {
+  public void putNoReturn(K key, V value, long createTimeInSecs, int customMaxTTISeconds, int customMaxTTLSeconds) {
     // custom tti ttl not supported
     put(key, value);
   }
 
   @Override
-  public V putIfAbsent(K key, V value, int createTimeInSecs, int customMaxTTISeconds, int customMaxTTLSeconds) {
+  public V putIfAbsent(K key, V value, long createTimeInSecs, int customMaxTTISeconds, int customMaxTTLSeconds) {
     throw new UnsupportedOperationException();
   }
